requirements.txt
torch
torchvision
numpy
opencv-python
cryptography
liboqs-python==0.14.0
